refactor, remove mocks
