+++ src/components/pages/History/HistoryPage.js
@@ -5,14 +5,13 @@ import { HistoryContainer, HistoryList } from "./HistoryStyle"
 import api from '../../../services/api.js';
 import { PageLoading } from "../../shared/Loadings.js";
 import { FaGhost, FaArrowDown, FaArrowUp } from "react-icons/fa";
-
+import { getDate, getHour, scrollToNextItem } from "./HistoryFunctions.js";
 export default function HistoryPage() {
     const { userData } = useContext(UserDataContext);
     const [purchaseHistory, setPurchaseHistory] = useState([])
     const navigate = useNavigate();
     const [pageLoading, setPageLoading] = useState(true);
     const listRefs = useRef(null);
-    const [actualView, setActualView] = useState(0);
 
     useEffect(() => {
         if (userData.userId) {
@@ -23,30 +22,14 @@ export default function HistoryPage() {
     },[])
 
     function renderCart() {
-        api.getPurchaseHistory("2035f9f7-0709-4711-9ff2-0874c20e85c1")
+        api.getPurchaseHistory(userData.token)
         .then(res => {
             setPurchaseHistory(res.data)
             listRefs.current = res.data.map(() => createRef())
-            console.log(listRefs)
 
             setPageLoading(false)
         })
         .catch (err => console.log(err))
-    }
-
-    function getDate (date) {
-        return (new Date(date).toLocaleString()).split(' ')[0]
-    }
-
-    function getHour (date) {
-        return (new Date(date).toLocaleString()).split(' ')[1]
-    }
-
-    function scrollToNextItem(dir, idx) {
-        window.scrollTo({
-            top: listRefs.current[dir === 'down' ? idx+1 : idx-1].current.offsetTop - 100,
-            behavior: 'smooth'
-        })
     }
 
     if (pageLoading) return <PageLoading />
@@ -54,7 +37,7 @@ export default function HistoryPage() {
         <HistoryContainer>
             {
             purchaseHistory.length === 0 ? <div className='empty'><FaGhost /> <span>Empty history</span></div> : 
-            <HistoryList>
+            <HistoryList >
                 {
                     purchaseHistory.map((item, idx) => (
                         <li ref={listRefs.current[idx]}>
@@ -81,10 +64,10 @@ export default function HistoryPage() {
                             </div>
 
                             {  idx === 0 ? '' :
-                                <div className='arrow-next up'><FaArrowUp  onClick={() => scrollToNextItem('up', idx)}/></div>
+                                <div className='arrow-next up'><FaArrowUp  onClick={() => scrollToNextItem(listRefs, 'up', idx)}/></div>
                             }
                             {  idx === purchaseHistory.length-1 ? '' :
-                                <div className='arrow-next down'><FaArrowDown  onClick={() => scrollToNextItem('down', idx)}/></div>
+                                <div className='arrow-next down'><FaArrowDown  onClick={() => scrollToNextItem(listRefs, 'down', idx)}/></div>
                             }
                         </li>
                     ))
